Update usb_images_storage.py
This program demonstrates the use of simple functions in Python. It prompts the user for the size of a USB drive (in Gigabytes or Gigs) and passes that value to different functions which calculate the number of images that can be stored on it in various image file formats (gif, jpeg, png and tiff).